Cu 861m6c6wm 20 create web security classes and jwt utility andrej mitrovi

+++ Tim26-ISS/src/main/java/com/example/demo/repository/RouteRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Route;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RouteRepository extends JpaRepository<Route, Integer> {
+public interface RouteRepository extends JpaRepository<Route, Long> {
 }
