+++ Tim26-ISS/src/main/java/com/example/demo/repository/VehicleTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.VehicleType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleTypeRepository extends JpaRepository<VehicleType, Integer> {
+public interface VehicleTypeRepository extends JpaRepository<VehicleType, Long> {
 }
